moved run fetching code to separate file
